requirements.txt
python-dotenv
google-generativeai
python-telegram-bot==13.15
aiohttp
beautifulsoup4